web/requirements_web.txt
streamlit==1.27.2
st-annotated-text==4.0.0
requests==2.31.0
python-dotenv==1.0.0